fix(wrapper): add missing awaits before page close
Closing the page at the end of the request
processing failed, because `await`. was missing
that waits for the processing to end. Therefore
the `close()` happened too early, before the
processing has even startet. Thus, no page was
available at that time.

Because the close() happened too early, the page
was not closed at the end of the request processing.