refactor(api): `PromiseLike` instead of `Promise`
Wherever possible `PromiseLike<>` is used instead
of `Promise<>`. It seems as if this is the preferred
way to avoid depending on a specific Promise library.

At least, that was what some commenters on various
blogs wrote.

It is a proper replacement and does not hurt. So, we
give it a try.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>