feat(api): introduce long running task queue
Driving the Ariba website takes a lot more
time than many callers expect from an API
request. Hence, the task is regarded as
a "long running task".

A `Task Manager` is introduced to handle such
tasks and execute the in-process but outside
of the request.

Because the main work of driving the Ariba
website is waiting for the headless browser
to perform the commands, there seems not much
to do for the node process but waiting.
Hence, no separate process is spawn or the task
otherwise processed by a separate service.

see:
* http://restalk-patterns.org/long-running-operation-polling.html
* https://medium.com/geekculture/rest-api-best-practices-decouple-long-running-tasks-from-http-request-processing-9fab2921ace8

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>