fix(wrapper): inject jQuery if not available
JQuery is used a lot to find HTML nodes.
Some pages have loaded JQuery, but some
have not. So, JQuery can be injected into
the page if necessary.

Because JQuery is loaded from the Ariba
website, this is possible. It can not be
loaded from arbitrary URLs because of
the CORS headers.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>