refactor(login): restart fresh login after some time
It seems as if the ARIBA web server has a problem with long running
session, running more than a day or two. To overcome its issue,
the session is destroyed (=cookies) and a fresh is started with a
fresh login after the login has been refreshed a couple of times.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>