refactor(login): check for enforced login page
Occasionally, the server enforces a re-login by providing a login page
instead of the expected purchase order page or invoice list page.
However, the dashboard is unaffected and displayed properly.
Hence, this was not detected. Now, almost each new page reload is
checked to see, whether a login page is being presented by the server.
In such a case, login is performed and the expected page is opened
again.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>