fix(invoice): support partially delivered orders
In case parts of an order has been rejected or revoked, the delivery
status of the order becomes "partially delivered".

Previous detection of status and mapping to constants did not support
this kind of status. Now, this is supported and invoices will be
created for such orders too.

There are more order states that lack support because no such case
has been detected yet to test with.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>