fix(invoice): adapt page check to website changes
ARIBA WebSite has changed and moved to a less table oriented design.
Thus the check for the open invoice page failed, because it expected
a table cell and a different class.

The new design uses a div-layer and an ID. Hence, the check es
even easier. The opened invoice search page can be detected now.

Signed-off-by: Nikolaus Rosenmayr <nr@firecoder.com>